drone22111/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "email-triage-openenv"
version = "1.0.0"
description = "A real-world task simulation for email triage compliant with OpenEnv."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0.0",
    "fastapi",
    "uvicorn",
    "openai",
    "python-dotenv",
    "gymnasium",
    "numpy",
    "flask",
    "flask-cors",
]

[project.optional-dependencies]
dev = ["pytest"]
